[CWS] extract trace pipe logger to different file
